aegis_back/requirements.txt
asgiref==3.8.1
blinker==1.4
certifi==2024.12.14
charset-normalizer==3.4.1
command-not-found==0.3
cryptography==3.4.8
dbus-python==1.2.18
distro==1.7.0
distro-info==1.1+ubuntu0.2
Django==5.1.5
django-cors-headers==4.6.0
django-extensions==3.2.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
httplib2==0.20.2
humanize==4.11.0
idna==3.10
importlib-metadata==4.6.4
jeepney==0.7.1
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
pillow==11.1.0
PyGObject==3.42.1
PyJWT==2.10.1
pyparsing==2.4.7
python-apt==2.4.0+ubuntu3
pytz==2022.1
PyYAML==5.4.1
requests==2.32.3
SecretStorage==3.3.1
six==1.16.0
sqlparse==0.5.3
systemd-python==234
typing_extensions==4.12.2
tzdata==2024.2
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==2.3.0
wadllib==1.3.6
zipp==1.0.0
